backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
supabase>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0